+++ src/components/Cart/index.tsx
@@ -1,9 +1,14 @@
 import * as Dialog from '@radix-ui/react-dialog';
 import { Handbag, X } from 'phosphor-react';
 import { useCart } from '../../contexts/CartContext';
-import { HeaderButtonContainer } from '../Header/styles';
 import { CartContent } from './CartContent';
-import { Close, Content, Description, Overlay } from './styles';
+import {
+   CartButtonContainer,
+   Close,
+   Content,
+   Description,
+   Overlay,
+} from './styles';
 
 export const Cart = () => {
    const { cartQuantity } = useCart();
@@ -13,10 +18,10 @@ export const Cart = () => {
    return (
       <Dialog.Root>
          <Dialog.Trigger asChild>
-            <HeaderButtonContainer>
+            <CartButtonContainer>
                <Handbag size={24} />
                {cartHasContent ? <span>{cartQuantity}</span> : <></>}
-            </HeaderButtonContainer>
+            </CartButtonContainer>
          </Dialog.Trigger>
 
          <Dialog.Portal>
