+++ src/components/Cart/styles.ts
@@ -11,6 +11,42 @@ const fadeOutAnimation = keyframes({
    '100%': { opacity: 0 },
 });
 
+export const CartButtonContainer = styled('button', {
+   width: '3rem',
+   height: '3rem',
+
+   border: 0,
+   borderRadius: 6,
+
+   background: '$gray800',
+   color: '$gray300',
+
+   cursor: 'pointer',
+
+   position: 'relative',
+
+   span: {
+      position: 'absolute',
+      width: '1.75rem',
+      height: '1.75rem',
+
+      top: 'calc(-1.25rem / 2)',
+      right: 'calc(-1.25rem / 2)',
+
+      background: '$green500',
+      color: '$gray100',
+
+      border: '3px solid $gray900',
+      borderRadius: '1000px',
+
+      display: 'flex',
+      alignItems: 'center',
+      justifyContent: 'center',
+
+      fontWeight: 700,
+   },
+});
+
 export const Overlay = styled(Dialog.Overlay, {
    position: 'fixed',
    inset: 0,
