+++ src/components/HeaderWithCart/index.tsx
@@ -2,9 +2,10 @@ import Image from 'next/image';
 import Link from 'next/link';
 import logoImg from '../../assets/logo.svg';
 import { Cart } from '../Cart';
+
 import { HeaderContainer } from './styles';
 
-export const Header = () => {
+export const HeaderWithCart = () => {
    return (
       <HeaderContainer>
          <Link href='/'>
