+++ src/components/HeaderWithCart/styles.ts
@@ -1,0 +1,12 @@
+import { styled } from '../../styles';
+
+export const HeaderContainer = styled('header', {
+   width: '100%',
+   maxWidth: 1180,
+
+   margin: '0 auto',
+   padding: '2rem 0',
+
+   display: 'flex',
+   justifyContent: 'space-between',
+});
